Implement contain for ray paths
When 'contain' is specified on a ray path, the distance along the
path is clamped so that the element lies entirely within the path.

When no offset-distance would place the element entirely within the
path, the path length is minimally increased such that an
offset-distance exists that places the element within the path.

Spec:
https://drafts.fxtf.org/motion-1/#valdef-offsetpath-ray

Design doc:
https://docs.google.com/document/d/1d2AYD2_SXRYH6Wua3pqFe5H3Pu0GTm2KD2P5Vi0swEg/