Clamp before adding epsilon
When an offset-path path string is not a closed path (i.e. does not
end with Z) and offset-distance is negative or greater than the path
length, the following all give the same point:
- pathElement.getPointAtLength(distance)
- pathElement.getPointAtLength(distance + epsilon)
- pathElement.getPointAtLength(distance - epsilon)

Then we don't obtain any direction information.

An example of the bug this causes is the green box in
https://ewilligers.github.io/motion-path-polyfill-examples/spec/offset-distance-unclosed.html

The green box should have been rotated: See EXAMPLE 5 / Figure 7 in
https://drafts.fxtf.org/motion-1/#offset-distance-property

We need to clamp to [0, path length] before adding epsilon / -epsilon.